src/test/java/mx/edu/utez/sistemaContratos/Categoria/Categoria.java
package mx.edu.utez.sistemaContratos.Categoria;

public class Categoria {
    private String nombre;
    private String descripcion;
    private boolean estado; // Activo o Inactivo

    public Categoria() {
    }

    public Categoria(String nombre, String descripcion, boolean estado) {
        this.nombre = nombre;
        this.descripcion = descripcion;
        this.estado = estado;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public String getDescripcion() {
        return descripcion;
    }

    public void setDescripcion(String descripcion) {
        this.descripcion = descripcion;
    }

    public boolean isEstado() {
        return estado;
    }

    public void setEstado(boolean estado) {
        this.estado = estado;
    }
}
